made a calculator

+++ javaplayground/src/main/java/com/nerunsmarts/Calculator.java
@@ -4,19 +4,19 @@ public class Calculator {
 
     public Calculator() {}
 
-    public int add(int a, int b) {
+    public float add(int a, int b) {
         return a + b;
     }
 
-    public int subtract(int a, int b) {
+    public float subtract(int a, int b) {
         return a - b;
     }
 
-    public int multiply(int a, int b) {
+    public float multiply(int a, int b) {
         return a * b;
     }
 
-    public int divide(int a, int b) throws Exception {
+    public float divide(int a, int b) throws Exception {
        if (b != 0) {
               return a / b;
          } else {
